requirements.txt
pandas
numpy
psycopg2-binary
mlflow
dagshub
joblib
scikit-learn==1.2.2
pywebio
flask
catboost
lightgbm
python-dotenv
joblib
-e .